Use child_process.fork to start ssr build
